Rename old parse/1 to build_resource_map/1; add parse/1:
parse/1 now triages the supplied result according to its content type
and whether it has the right unique ID field for newspapers. If the
result is either not a newspaper article or is and has the right
field, passes the result to build_resource_map/1. If not, returns an
empty map.